Restructure 'math' package and adjust imports
