+++ src/main/java/bingo/math/operators/impl/Add.java
@@ -1,6 +1,6 @@
-package bingo.math.operator.impl;
+package bingo.math.operators.impl;
 
-import bingo.math.operator.Operator;
+import bingo.math.operators.Operator;
 
 public final class Add implements Operator {
     public static final Operator OPERATOR = new Add();
