+++ src/main/java/bingo/math/terms/impl/Addition.java
@@ -1,7 +1,7 @@
-package bingo.math.impl;
+package bingo.math.terms.impl;
 
-import bingo.math.Term;
-import bingo.math.operator.impl.Add;
+import bingo.math.operators.impl.Add;
+import bingo.math.terms.Term;
 
 public final class Addition extends TermWithOperator {
