+++ src/main/java/bingo/math/terms/impl/Equation.java
@@ -1,6 +1,6 @@
-package bingo.math.impl;
+package bingo.math.terms.impl;
 
-import bingo.math.Term;
+import bingo.math.terms.Term;
 
 public final class Equation extends DelegateTerm {
