+++ src/main/java/bingo/math/terms/impl/Literal.java
@@ -1,6 +1,6 @@
-package bingo.math.impl;
+package bingo.math.terms.impl;
 
-import bingo.math.Term;
+import bingo.math.terms.Term;
 
 public final class Literal implements Term {
     private final double value;
