+++ src/main/java/bingo/ribbons/Ribbon.java
@@ -1,8 +1,8 @@
 package bingo.ribbons;
 
-import bingo.math.impl.LabeledTerm;
-import bingo.math.impl.Literal;
-import bingo.math.impl.TermWithPoints;
+import bingo.math.terms.impl.LabeledTerm;
+import bingo.math.terms.impl.Literal;
+import bingo.math.terms.impl.TermWithPoints;
 import bingo.ribbons.overrides.PointValueOverride;
 import bingo.ships.MainArmamentType;
